// packages/panel-vendure/src/plugins/chasqui-plugin/productores/generated-types2.ts
export type Maybe<T> = T;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the `date-time` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
  DateTime: any;
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: any;
  /** The `Upload` scalar type represents a file upload. */
  Upload: any;
};
// ...
export type Asset = Node & {
  __typename?: 'Asset';
  tags: Array<Tag>;
  id: Scalars['ID'];
  createdAt: Scalars['DateTime'];
  updatedAt: Scalars['DateTime'];
  name: Scalars['String'];
  type: AssetType;
  fileSize: Scalars['Int'];
  mimeType: Scalars['String'];
  width: Scalars['Int'];
  height: Scalars['Int'];
  source: Scalars['String'];
  preview: Scalars['String'];
  focalPoint: Maybe<Coordinate>;
  customFields: Maybe<Scalars['JSON']>;
};
// ...
export enum AssetType {
  IMAGE = 'IMAGE',
  VIDEO = 'VIDEO',
  BINARY = 'BINARY'
}
// ...
export type Coordinate = {
  __typename?: 'Coordinate';
  x: Scalars['Float'];
  y: Scalars['Float'];
};
// ...
export type Node = {
  id: Scalars['ID'];
};
// ...
export type PaginatedList = {
  items: Array<Node>;
  totalItems: Scalars['Int'];
};
// ...
export type Product = Node & {
  __typename?: 'Product';
  enabled: Scalars['Boolean'];
  channels: Array<Channel>;
  id: Scalars['ID'];
  createdAt: Scalars['DateTime'];
  updatedAt: Scalars['DateTime'];
  languageCode: LanguageCode;
  name: Scalars['String'];
  slug: Scalars['String'];
  description: Scalars['String'];
  featuredAsset: Maybe<Asset>;
  assets: Array<Asset>;
  variants: Array<ProductVariant>;
  optionGroups: Array<ProductOptionGroup>;
  facetValues: Array<FacetValue>;
  translations: Array<ProductTranslation>;
  collections: Array<Collection>;
  customFields: Maybe<ProductCustomFields>;
};
// ...
export type ProductList = PaginatedList & {
  __typename?: 'ProductList';
  items: Array<Product>;
  totalItems: Scalars['Int'];
};
// ...
export type Productor = Node & {
  __typename?: 'Productor';
  enabled: Maybe<Scalars['Boolean']>;
  name: Scalars['String'];
  slug: Scalars['String'];
  description: Maybe<Scalars['String']>;
  descriptionOffered: Maybe<Scalars['String']>;
  createdAt: Scalars['DateTime'];
  updatedAt: Scalars['DateTime'];
  id: Scalars['ID'];
  products: Maybe<ProductList>;
  sellos: Maybe<Array<Maybe<Scalars['String']>>>;
  logo: Maybe<Asset>;
  banner: Maybe<Asset>;
  pais: Maybe<Scalars['String']>;
  provincia: Maybe<Scalars['String']>;
  localidad: Maybe<Scalars['String']>;
  direccion: Maybe<Scalars['String']>;
  webUrl: Maybe<Scalars['String']>;
  email: Maybe<Scalars['String']>;
  linksRRSS: Maybe<Array<Maybe<Scalars['String']>>>;
};


export type ProductorProductsArgs = {
  options?: Maybe<ProductListOptions>;
};

export type ProductorAddInput = {
  enabled: Maybe<Scalars['Boolean']>;
  name: Scalars['String'];
  slug: Scalars['String'];
  description: Maybe<Scalars['String']>;
  descriptionOffered: Maybe<Scalars['String']>;
  sellos: Maybe<Array<Maybe<Scalars['String']>>>;
  logoId: Maybe<Scalars['ID']>;
  bannerId: Maybe<Scalars['ID']>;
  pais: Maybe<Scalars['String']>;
  provincia: Maybe<Scalars['String']>;
  localidad: Maybe<Scalars['String']>;
  direccion: Maybe<Scalars['String']>;
  webUrl: Maybe<Scalars['String']>;
  email: Maybe<Scalars['String']>;
  linksRRSS: Maybe<Array<Maybe<Scalars['String']>>>;
};
// ...
export type ProductorUpdateInput = {
  enabled: Maybe<Scalars['Boolean']>;
  id: Scalars['ID'];
  name: Scalars['String'];
  slug: Scalars['String'];
  description: Maybe<Scalars['String']>;
  descriptionOffered: Maybe<Scalars['String']>;
  sellos: Maybe<Array<Maybe<Scalars['String']>>>;
  logoId: Maybe<Scalars['ID']>;
  bannerId: Maybe<Scalars['ID']>;
  pais: Maybe<Scalars['String']>;
  provincia: Maybe<Scalars['String']>;
  localidad: Maybe<Scalars['String']>;
  direccion: Maybe<Scalars['String']>;
  webUrl: Maybe<Scalars['String']>;
  email: Maybe<Scalars['String']>;
  linksRRSS: Maybe<Array<Maybe<Scalars['String']>>>;
};
// ...
export type Tag = Node & {
  __typename?: 'Tag';
  id: Scalars['ID'];
  createdAt: Scalars['DateTime'];
  updatedAt: Scalars['DateTime'];
  value: Scalars['String'];
};